* refactoring - moved constants from plasmaSwitchboard to own class (all 232 ;) - moved remoteProxy-Methods to httpRemoteProxyConfig, better names - removed some unnecessary code (else-statements) * formatting (correct indentation) * minor bugfixes (due to findbugs.sf.net) * hopefully fixed "missing quote" (announcing StringParts as UTF-8)
git-svn-id: https://svn.berlios.de/svnroot/repos/yacy/trunk@5031 6c8d7289-2bf4-0310-a012-ef5d649a1542

+++ source/de/anomic/plasma/plasmaDHTFlush.java
@@ -110,7 +110,7 @@ public class plasmaDHTFlush extends Thread {
         public int getTransferedEntrySpeed() {
             long transferTime = System.currentTimeMillis() - this.startingTime;
             if (transferTime <= 0) transferTime = 1;
-            return (int) ((1000L * (long) this.transferedEntryCount) / transferTime);
+            return (int) ((1000L * this.transferedEntryCount) / transferTime);
         }
         
         public yacySeed getSeed() {
